refactor: move builders randomise to general hook
