+++ src/app/backing/components/builder-card/BuildersContainer/BuildersContainer.tsx
@@ -2,13 +2,16 @@ import { BuilderCardContainer } from '@/app/backing/components/builder-card/Buil
 import { useGetEstimatedBackersRewardsPct } from '@/app/collective-rewards/shared/hooks/useGetEstimatedBackersRewardsPct'
 import { Button } from '@/components/Button'
 import { LoadingSpinner } from '@/components/LoadingSpinner'
-import { FC, useState, useRef } from 'react'
+import { FC } from 'react'
+import { useRouter } from 'next/navigation'
+import { useShuffledArray } from './useShuffledArray'
+import { Builder, BackerRewardsConfig } from '@/app/collective-rewards/types'
 
 export const BuildersContainer: FC = () => {
   const { data: builders, isLoading, error } = useGetEstimatedBackersRewardsPct()
   const [showAll, setShowAll] = useState(false)
-  // Generate a random seed once per session
-  const seedRef = useRef<number>(Math.floor(Math.random() * 1_000_000_000))
+  const router = useRouter()
+  const shuffledBuilders = useShuffledArray<Builder>(builders)
 
   if (isLoading) {
     return <LoadingSpinner />
@@ -18,21 +21,23 @@ export const BuildersContainer: FC = () => {
     return <div>Error loading builders: {error.message}</div>
   }
 
-  const transformedBuilders = builders
-    ? seededShuffle(builders, seedRef.current).map(builder => ({
-        ...builder,
-        backerRewardPercentage: {
-          active: builder.rewardPercentage?.current ?? 0n,
-          previous: builder.rewardPercentage?.current ?? 0n,
-          next: builder.rewardPercentage?.next ?? 0n,
-          cooldown: builder.rewardPercentage?.cooldownEndTime ?? 0n,
-        },
-      }))
-    : []
-
   if (!transformedBuilders || transformedBuilders.length === 0) {
     return <div>No builders found</div>
   }
+
+  const transformedBuilders = shuffledBuilders.map((builder: Builder) => {
+    const backerRewardPercentage: BackerRewardsConfig = {
+      active: builder.backerRewardPercentage?.active ?? 0n,
+      previous: builder.backerRewardPercentage?.previous ?? 0n,
+      next: builder.backerRewardPercentage?.next ?? 0n,
+      cooldown: builder.backerRewardPercentage?.cooldown ?? 0n,
+    }
+
+    return {
+      ...builder,
+      backerRewardPercentage,
+    }
+  })
 
   const spotlightBuilders = 4
   const visibleBuilders = showAll ? transformedBuilders : transformedBuilders.slice(0, spotlightBuilders)
@@ -51,25 +56,4 @@ export const BuildersContainer: FC = () => {
       </div>
     </>
   )
-}
-
-// Seeded shuffle
-function seededShuffle<T>(array: T[], seed: number): T[] {
-  const arr = [...array]
-  const random = mulberry32(seed)
-  for (let i = arr.length - 1; i > 0; i--) {
-    const j = Math.floor(random() * (i + 1))
-    ;[arr[i], arr[j]] = [arr[j], arr[i]]
-  }
-  return arr
-}
-
-// Simple seeded random number generator (Mulberry32)
-function mulberry32(seed: number) {
-  return function () {
-    let t = (seed += 0x6d2b79f5)
-    t = Math.imul(t ^ (t >>> 15), t | 1)
-    t ^= t + Math.imul(t ^ (t >>> 7), t | 61)
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296
-  }
 }
